Rename SharezonePlusFeatureGuard file.
